+++ src/Helpers/MiscHelper.php
@@ -22,13 +22,13 @@ class MiscHelper
     public function getGeneralHelpString(string $type) {
         $helpStrings = [
             'peasants' => 'Peasants are the non-military part of your population. They pay taxes and get drafted into military service.',
-            'employment' => 'Each employed peasant pays 2.7p/h in taxes.',
+            'employment' => 'Each employed peasant pays 2.7 platinum per hour in taxes.',
             'networth' => 'Used to determine power of a dominion.<br>Buildings, land, and units give networth.',
-            'prestige' => 'Gained via invasion.<br>Increases offensive power, maximum population, and food production.',
+            'prestige' => 'Gained via invasion.<br>Every point of prestige increases offensive power, maximum population, and food production by 0.01%.',
             'morale' => 'Morale below 100% gives a defensive penalty.<br>Morale is lowered by exploring and invading.',
-            'infamy' => 'Gained via war operations.<br>Increases platinum, gem, lumber, and ore production.',
-            'spy_mastery' => 'Gained and lost via war operations.<br>Used for spy rankings.',
-            'wizard_mastery' => 'Gained and lost via war operations.<br>Used for wizard rankings.',
+            'infamy' => 'Gained via war operations.<br>Increases platinum, gem, lumber, and ore production. Decays over time.',
+            'spy_mastery' => 'Gained and lost via war operations.<br>Every 100 mastery adds 50 to your minimum infamy.',
+            'wizard_mastery' => 'Gained and lost via war operations.<br>Every 100 mastery adds 50 to your minimum infamy.',
             'spy_resilience' => 'Gained by victims of war operations.<br>Reduces damage taken from spy ops.',
             'wizard_resilience' => 'Gained by victims of war operations.<br>Reduces damage taken from wizard ops.',
         ];